refactor($step4): refactor LottoGamePlayer, Money
Breaks ProfitCalculator was romoved (use Money instead)
