+++ src/main/java/lotto/domain/Money.java
@@ -1,9 +1,6 @@
 package lotto.domain;
 
 import lotto.common.ErrorCode;
-import lotto.common.WinningType;
-
-import java.util.List;
 
 public class Money {
     private static final int LOTTO_TICKET_PRICE = 1000;
@@ -21,7 +18,6 @@ public class Money {
     }
 
     public double calculateProfit(long prizeSum) {
-        ProfitCalculator profitCalculator = new ProfitCalculator();
-        return profitCalculator.calculate(this.money, prizeSum);
+        return Math.floor(((double) prizeSum/money)*100)/100.0;
     }
 }
