+++ src/main/java/lotto/view/ResultView.java
@@ -13,14 +13,6 @@ public class ResultView {
         System.out.println(lottoTicketsText);
     }
 
-    public void printResultStatistics(int[] matchCountResults) {
-        System.out.println("\n당첨 통계\n---------");
-        for(int i=3; i<matchCountResults.length; i++) {
-            System.out.println(i + "개 일치 (" + WinningType.of(i).getPrize() + ")-"
-                    + matchCountResults[i] +"개");
-        }
-    }
-
     public void printResultStatistics(Map<WinningType, Integer> gameResult) {
         System.out.println("\n당첨 통계\n---------");
         for(WinningType winningType : gameResult.keySet()) {
