+++ src/test/java/lotto/domain/LottoTicketTest.java
@@ -16,7 +16,7 @@ public class LottoTicketTest {
 
         // 6개 모두 일치 시에는 1등
         LottoTicket userFirstLottoTicket = new LottoTicket("1,2,3,4,5,6");
-        winningType = userFirstLottoTicket.getWinningType(winningLottoTicket);
+        winningType = userFirstLottoTicket.getWinningType(winningLottoTicket, 10);
         assertThat(winningType).isEqualTo(WinningType.FIRST);
 
         //2등은 보너스 번호가 포함되어있고, 5개 로또번호가 일치하는 경우이다.
@@ -29,12 +29,12 @@ public class LottoTicketTest {
 
         //5등은 3개 로또번호가 일치하는 경우이다.
         LottoTicket userForthLottoTicket = new LottoTicket("1,2,3,14,15,16");
-        winningType = userForthLottoTicket.getWinningType(winningLottoTicket);
+        winningType = userForthLottoTicket.getWinningType(winningLottoTicket, 10);
         assertThat(winningType).isEqualTo(WinningType.FIFTH);
 
         //8등은 0개 로또번호가 일치하는 경우이다.
         LottoTicket userSeventhLottoNumbers = new LottoTicket("11,12,13,14,15,16");
-        winningType = userSeventhLottoNumbers.getWinningType(winningLottoTicket);
+        winningType = userSeventhLottoNumbers.getWinningType(winningLottoTicket, 10);
         assertThat(winningType).isEqualTo(WinningType.EIGHTH);
     }
 }
